+++ CHANGELOG.md
@@ -7,8 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- New features - Assertions
+- Added `expect()`
+- Added:
+  - `toBeDefined()`
+  - `toHaveLength()`
+
 ### Changed
-- Update core
+- Updated core
+- Updated documentation
 
 ## 1.3.0 - 2024-01-11
 ### Added