+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `.toBeFalsy()`
   - `.toBeTruthy()`
   - `.toEqual()`
+  - `.notToEqual()`
 
 ### Changed
 - Updated core